requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-allauth>=0.57.0
requests>=2.31.0
requests-oauthlib>=1.3.0
PyJWT>=2.8.0
cryptography>=41.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0
whitenoise>=6.0
dj-database-url>=2.0
python-dotenv>=1.0
